firmware: Change firmware/host to firmware module
