+++ src/firmware/linux/host/types.rs
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::certs::sev;
-use crate::firmware::host::types::{
+use crate::firmware::linux::_4K_PAGE;
+use crate::firmware::types::{
     CertTableEntry as UapiCertTableEntry, SnpCertError, SnpExtConfig, UserApiError,
 };
-use crate::firmware::linux::_4K_PAGE;
 use crate::Version;
 
 use std::marker::PhantomData;
