+++ src/firmware/types.rs
